Use local MediaComposition for Live urns
To avoid geoblock from only the media composition
