--- pillarbox-core-business/src/test/assets/media-composition.json
@@ -1,147 +1,0 @@
-{
-    "chapterUrn": "urn:rts:audio:3262363",
-    "episode": {
-        "id": "3262367",
-        "title": "Couleur 3 en direct",
-        "publishedDate": "2011-07-11T14:20:07+02:00",
-        "imageUrl": "https://www.rts.ch/2020/05/18/14/20/11333286.image/16x9",
-        "imageTitle": "Chaîne Couleur 3"
-    },
-    "show": {
-        "id": "3262370",
-        "vendor": "RTS",
-        "transmission": "RADIO",
-        "urn": "urn:rts:show:radio:3262370",
-        "title": "Couleur 3 en direct",
-        "imageUrl": "https://www.rts.ch/2020/05/18/14/20/11333286.image/16x9",
-        "imageTitle": "Chaîne Couleur 3",
-        "bannerImageUrl": "https://www.rts.ch/2020/05/18/14/20/11333286.image/3x1",
-        "posterImageUrl": "https://ws.srf.ch/asset/image/audio/e0322b37-5697-474d-93ac-19a4044a6a24/POSTER.jpg",
-        "posterImageIsFallbackUrl": true,
-        "primaryChannelId": "8ceb28d9b3f1dd876d1df1780f908578cbefc3d7",
-        "primaryChannelUrn": "urn:rts:channel:radio:8ceb28d9b3f1dd876d1df1780f908578cbefc3d7",
-        "audioDescriptionAvailable": false,
-        "subtitlesAvailable": false,
-        "multiAudioLanguagesAvailable": false,
-        "allowIndexing": false
-    },
-    "channel": {
-        "id": "8ceb28d9b3f1dd876d1df1780f908578cbefc3d7",
-        "vendor": "RTS",
-        "urn": "urn:rts:channel:radio:8ceb28d9b3f1dd876d1df1780f908578cbefc3d7",
-        "title": "Couleur 3",
-        "imageUrl": "https://www.rts.ch/2020/05/18/14/20/11333286.image/16x9",
-        "imageTitle": "Chaîne Couleur 3",
-        "transmission": "RADIO"
-    },
-    "chapterList": [
-        {
-            "id": "3262363",
-            "mediaType": "AUDIO",
-            "vendor": "RTS",
-            "urn": "urn:rts:audio:3262363",
-            "title": "Couleur 3 en direct",
-            "imageUrl": "https://www.rts.ch/2020/05/18/14/20/11333286.image/16x9",
-            "imageTitle": "Chaîne Couleur 3",
-            "type": "LIVESTREAM",
-            "date": "2011-07-11T14:20:07+02:00",
-            "duration": 0,
-            "playableAbroad": true,
-            "displayable": true,
-            "position": 0,
-            "noEmbed": false,
-            "analyticsMetadata": {
-                "media_segment": "Livestream",
-                "media_type": "Audio",
-                "media_segment_id": "3262363",
-                "media_episode_length": "0",
-                "media_segment_length": "0",
-                "media_number_of_segment_selected": "1",
-                "media_number_of_segments_total": "1",
-                "media_duration_category": "infinit.livestream",
-                "media_is_geoblocked": "false",
-                "media_is_web_only": "false",
-                "media_production_source": "produced.for.broadcasting",
-                "media_urn": "urn:rts:audio:3262363"
-            },
-            "fullLengthMarkIn": 0,
-            "fullLengthMarkOut": 0,
-            "resourceList": [
-                {
-                    "url": "http://lsaplus.swisstxt.ch/audio/couleur3_96.stream/playlist.m3u8?",
-                    "quality": "HD",
-                    "protocol": "HLS-DVR",
-                    "encoding": "H264",
-                    "mimeType": "application/x-mpegURL",
-                    "presentation": "DEFAULT",
-                    "streaming": "HLS",
-                    "dvr": true,
-                    "live": true,
-                    "mediaContainer": "MPEG2_TS",
-                    "audioCodec": "AAC",
-                    "videoCodec": "NONE",
-                    "tokenType": "NONE",
-                    "analyticsMetadata": {
-                        "media_streaming_quality": "HD",
-                        "media_special_format": "DEFAULT",
-                        "media_url": "http://lsaplus.swisstxt.ch/audio/couleur3_96.stream/playlist.m3u8?"
-                    },
-                    "streamOffset": 55000
-                }
-            ]
-        }
-    ],
-    "analyticsData": {
-        "srg_pr_id": "3262367",
-        "srg_plid": "3262370",
-        "ns_st_pl": "Livestream",
-        "ns_st_pr": "Couleur 3 en direct",
-        "ns_st_dt": "2011-07-11",
-        "ns_st_ddt": "2011-07-11",
-        "ns_st_tdt": "2011-07-11",
-        "ns_st_tm": "14:20:07",
-        "ns_st_tep": "*null",
-        "ns_st_li": "1",
-        "ns_st_stc": "0867",
-        "ns_st_st": "Couleur 3",
-        "ns_st_tpr": "11562086",
-        "ns_st_en": "*null",
-        "ns_st_ge": "*null",
-        "ns_st_ia": "*null",
-        "ns_st_ce": "1",
-        "ns_st_cdm": "to",
-        "ns_st_cmt": "fc",
-        "srg_unit": "RTS",
-        "srg_c1": "live",
-        "srg_c2": "rts.ch_audio_couleur3",
-        "srg_c3": "COULEUR 3",
-        "srg_aod_prid": "3262367"
-    },
-    "analyticsMetadata": {
-        "media_episode_id": "3262367",
-        "media_show_id": "11562086",
-        "media_show": "Oui Mais Non",
-        "media_episode": "Couleur 3 en direct",
-        "media_is_livestream": "true",
-        "media_full_length": "full",
-        "media_enterprise_units": "RTS",
-        "media_joker1": "live",
-        "media_joker2": "rts.ch_audio_couleur3",
-        "media_joker3": "COULEUR 3",
-        "media_is_web_only": "false",
-        "media_production_source": "produced.for.broadcasting",
-        "media_thumbnail": "https://www.rts.ch/2020/05/18/14/20/11333286.image/16x9/scale/width/344",
-        "media_publication_date": "2011-07-11",
-        "media_publication_time": "14:20:07",
-        "media_publication_datetime": "2011-07-11T14:20:07+02:00",
-        "media_tv_date": "2011-07-11",
-        "media_tv_time": "14:20:07",
-        "media_tv_datetime": "2011-07-11T14:20:07+02:00",
-        "media_content_group": "Couleur 3",
-        "media_channel_id": "8ceb28d9b3f1dd876d1df1780f908578cbefc3d7",
-        "media_channel_cs": "0867",
-        "media_channel_name": "Couleur 3",
-        "media_since_publication_d": "4322",
-        "media_since_publication_h": "103747"
-    }
-}
